PT's output TSVs organization improvements
Pandas version now zfills the index and will put a dash in front of the parent file columns, meaning Terra will sort those two columns at the front. This is very useful when dealing with lots of child files so the user doesn't have to scroll all over the place to find out what the parent file extension is.